feat(shared-types): add omit and numberfy generics
